Split JS Wrapper (#2306)
