+++ .daily_canary
@@ -1,1 +1,1 @@
-Happy daily trigger :))
+Happy daily trigger :)))
