+++ CMakeLists.txt
@@ -64,8 +64,8 @@ if("sgx" IN_LIST COMPILE_TARGETS)
   # enclave version
   add_library(
     ccf.enclave STATIC
-    ${CCF_DIR}/src/enclave/main.cpp ${CCF_DIR}/src/enclave/thread_local.cpp
-    ${CCF_GENERATED_DIR}/ccf_t.cpp
+    ${CCF_DIR}/src/enclave/main.cpp ${CCF_DIR}/src/js/wrap.cpp
+    ${CCF_DIR}/src/enclave/thread_local.cpp ${CCF_GENERATED_DIR}/ccf_t.cpp
   )
 
   target_compile_definitions(
@@ -86,6 +86,7 @@ if("sgx" IN_LIST COMPILE_TARGETS)
     ccf.enclave
     PUBLIC ${OE_TARGET_ENCLAVE_AND_STD}
            -lgcc
+           quickjs.enclave
            ccfcrypto.enclave
            http_parser.enclave
            lua.enclave
@@ -111,8 +112,9 @@ endif()
 if("virtual" IN_LIST COMPILE_TARGETS)
   # virtual version
   add_library(
-    ccf.virtual STATIC ${CCF_DIR}/src/enclave/main.cpp
-                       ${CCF_DIR}/src/enclave/thread_local.cpp
+    ccf.virtual STATIC
+    ${CCF_DIR}/src/enclave/main.cpp ${CCF_DIR}/src/js/wrap.cpp
+    ${CCF_DIR}/src/enclave/thread_local.cpp
   )
 
   target_compile_definitions(
@@ -137,6 +139,7 @@ if("virtual" IN_LIST COMPILE_TARGETS)
            ccfcrypto.host
            http_parser.host
            lua.host
+           quickjs.host
            aft.virtual
            sss.host
            openenclave::oehost
@@ -361,12 +364,12 @@ if(BUILD_TESTS)
     target_link_libraries(http_test PRIVATE http_parser.host)
 
     add_unit_test(
-      frontend_test
+      frontend_test ${CMAKE_CURRENT_SOURCE_DIR}/src/js/wrap.cpp
       ${CMAKE_CURRENT_SOURCE_DIR}/src/node/rpc/test/frontend_test.cpp
     )
     target_link_libraries(
       frontend_test PRIVATE ${CMAKE_THREAD_LIBS_INIT} lua.host http_parser.host
-                            sss.host
+                            sss.host quickjs.host
     )
 
     add_unit_test(
@@ -375,12 +378,12 @@ if(BUILD_TESTS)
     )
 
     add_unit_test(
-      member_voting_test
+      member_voting_test ${CMAKE_CURRENT_SOURCE_DIR}/src/js/wrap.cpp
       ${CMAKE_CURRENT_SOURCE_DIR}/src/node/rpc/test/member_voting_test.cpp
     )
     target_link_libraries(
       member_voting_test PRIVATE ${CMAKE_THREAD_LIBS_INIT} lua.host
-                                 http_parser.host sss.host
+                                 http_parser.host sss.host quickjs.host
     )
 
     set_tests_properties(
@@ -390,12 +393,12 @@ if(BUILD_TESTS)
     )
 
     add_unit_test(
-      proposal_id_test
+      proposal_id_test ${CMAKE_CURRENT_SOURCE_DIR}/src/js/wrap.cpp
       ${CMAKE_CURRENT_SOURCE_DIR}/src/node/rpc/test/proposal_id_test.cpp
     )
     target_link_libraries(
       proposal_id_test PRIVATE ${CMAKE_THREAD_LIBS_INIT} lua.host
-                               http_parser.host sss.host
+                               http_parser.host sss.host quickjs.host
     )
 
     add_unit_test(
